Moving blogs to content collection
https://docs.astro.build/en/tutorial/6-islands/4/

[tag].astro
This is a dynamic route. It matches a single segment, like /posts/javascript or /posts/react. The value (e.g., javascript) is available as a parameter (tag).

[...slug].astro
The ... makes this a catch-all route (sometimes called a "splat" or "rest" parameter). It matches multiple segments, like /posts/2024/06/01/my-post.
The matched segments are available as an array (slug), e.g. ['2024', '06', '01', 'my-post'].